Add 0.6.11 conditions at pt-BR content_actorconditions.xml
The already existing actor's condition from the Portugal translation
are OK for Brazilian Portuguese, so keep them as-is. As "Blackwater"
is with a capital letter (and the corresponding translations that will
happen on other files will also use captal letters), it made sense
to put its translation ("Águas Negras") also in uppercase.

It is probably fine to use those conditions for Portugal's translation
as well, although I'm not 100% sure if they would call those conditions
the same way there.