Remove usage of Math functions from Coord and StatsModifierTraits.
Math.abs method call is not needed in Coord class. Function calls take
extra time and extra stack memory. Math.pow function is used in
StatsModifierTraits class. Is it possible formulas to be rewritten is
such way that Math.pow not to be used? In the case of calculateUseCost()
method it is possible by reorganize the order of the calculations. Some
test examples can be seen here:
https://drive.google.com/file/d/0Bz5a6NKy7qNiZzM3QTNweFV5MTA/edit?usp=sharing