Add auto-attack and improve confirm attack
This patch allows the player to specify in their preferences a
difficulty threshold for attack confirmations.  Thus, if you specify
'Normal', you will be prompted when attacking monsters of normal or
higher difficulty, but not easy or very easy. The options 'always' and
'never' implement the old boolean behavior.

Second, this adds an auto-attack setting that tells the game engine to
attempt to attack monsters as soon as you move next to them.  In the
preference interface, the player specifies the most difficult encounter
that they want to auto-attack.  Prior to auto-attacking, an agressive
monster will still get a chance to attack first. This modification only
acts after the player has moved; if a player is stationary and a
monster moves near them, they still must manually attack.